:recycle: Refactor DPoS unlocks to PoS unlocks in tests
